+++ src/components/LoginForm/LoginForm.tsx
@@ -1,22 +1,12 @@
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { useForm } from "react-hook-form";
-import * as S from "./LoginForm.style";
-import LoginInput from "./LoginInput/LoginInput";
-import LoginErrorModal from "./LoginErrorModal/LoginErrorModal";
-import handleQueryLogin from "./api/handleQueryLogin";
-import { handleHTTPResponseError } from "../../utils/auth/httpResponseUtils";
-import ROUTES from "../../constants/routes";
-
-import Cookies from "universal-cookie";
+import { useState } from 'react';
+import { useForm } from 'react-hook-form';
+import * as S from './LoginForm.style';
+import LoginInput from './LoginInput/LoginInput';
+import LoginErrorModal from './LoginErrorModal/LoginErrorModal';
+import { useLoginQuery } from './Login-query/Login.query';
 
 const LoginForm = () => {
-  const [serverAuthError, setServerAuthError] = useState("");
-  const navigate = useNavigate();
-  const cookies = new Cookies();
-
-  const expiresTime = new Date();
-  expiresTime.setHours(expiresTime.getHours() + 1);
+  const [serverAuthError, setServerAuthError] = useState('');
 
   const {
     register,
@@ -24,24 +14,11 @@ const LoginForm = () => {
     formState: { errors },
   } = useForm();
 
-  const { mutate: login } = handleQueryLogin({
-    onSuccess: (res: any) => {
-      cookies.set("access_token", res.accessToken, {
-        path: "/",
-        expires: expiresTime,
-      });
-
-      navigate(`${ROUTES.ACCOUNTS}`);
-    },
-
-    onError: (res) => {
-      setServerAuthError(handleHTTPResponseError(res));
-    },
-  });
+  const onSubmitMutate = useLoginQuery(setServerAuthError);
 
   return (
     <S.Container>
-      <S.Form onSubmit={handleSubmit((data: any) => login(data))}>
+      <S.Form onSubmit={handleSubmit((data: any) => onSubmitMutate(data))}>
         <S.Logo
           src="https://platum.kr/wp-content/uploads/2021/03/de.jpg"
           alt="디셈버앤컴퍼니"
