--- src/components/LoginForm/api/handleQueryLogin.ts
@@ -1,8 +1,0 @@
-import AuthAPI from "../../../libs/api/auth";
-import { useMutation, UseMutationOptions } from "@tanstack/react-query";
-
-const handleLogin = (options?: UseMutationOptions) => {
-  return useMutation(AuthAPI.login, options);
-};
-
-export default handleLogin;
